bee-check: --deps: provide detailed information about files and dependencies
if option --deps|-d is used we now provide detailed information about all files,
directories and symlinks provided by the package.

known dependencies to libraries or binaries are dicovered automatically and are
displayed for each file.

this prepares information generation mechanisms for a bee dependency tool.

the summarized view of package denpendencies (old behavior of -d) is removed
completely for now.

the output format may change in later releases.