beedep: first alpha version of teh bee dependency tool
This code was created by Lucas Schwass during a internship at the
Max Planck Institute for Molecular Genetics in 2011.

Parts of this code were created by Matthias Ruester and Marius Tolzmann.

The original history was squashed here and the author was set to Lucas,
because he wrote 90% of the code in this commit.

have fun 8)

not yet fully integrated into bee.