config-lib: add local config file
the current directory and all parents are searched for a file called .beerc
(BEE_LOCAL_CONFIG_FILENAME) and if found this file is read as a config file.

currently only the first file found is read.

after reading this file the user and system config is read.

this gives the possibility to define directory based configuration options.