beegetopt: improved ambiguity handling aka fix false ambiguity complaints
beegetopt --option update/upgrade/u -- --up complained about an ambiguous option.

Now the option is considered ambiguous if the returned value differs.

so beegetopt --option update/upgrade/u -- --up will return:

--update -- as expected.

Since this is fixed in bee_getopt.c this is also true for calls to the
bee_getop*() c-functions.

Only the first matching option is returned.

This implies that both options are of the same type. If they are not of the same type
undefined behavior may occur.