beesh: start guessing patch striplevels with 1 then 0
if a patch only contains new files from e.g. a git format-patch, the guess of
strip level 0 will succeed. This should be wrong most of the time, because
those files will be created in a subdirectory 'b' ..

So we now start guessing with striplevel 1 followed by 0 instead of starting
by 0 followed by 1 ..