bee/beegetopt: fix major option handling bug
we need to preserve the location of '--' option ending when pre parsing
commandline arguments in bee.

so beegetopt was changed to be able to do so if the new option
--keep-option-end is used.

there was an API-breaking change in bee_getopt() since flags are now
set in optctl and not via the third argument of (bee_getopt()).

since we also need the new flag in lower option parsing this was
the cleanest way to do so at the moment.

since this is new nobody should get hurt after we release this 8)

i'm really sorry for changing the API 8)