bee-download: git: use time of last git commit when no tag is found
when cloning git repositories and creating packages use the unix timestamp
of the last commit when creating version information in repositories that
do not have tags.

this makes it possible to better sort those archives since just adding
the commit-sha1 doesn't help bee sorting those versions.

the commit timestamp will be added as pkgextraversion prefixed with t
(e.g. PKGEXTRAVERION=_t1234567_deadbeef). the commit-sha1 will still be
added to make is possible to identify the commit the archive was built from.

pkgversion will still be 0.