bee_tree: add support for uniqueness of tree nodes
to enable dupe prevention set one of the flags below:

BEE_TREE_FLAG_UNIQUE - unique keys
   do not insert dupes

BEE_TREE_FLAG_UNIQUE_DATA - unique keys and data
   do not insert dupes - if key is already in tree
   also check data. tree->compare_data() must be set.
   data comparison is only used to prevent dupes and is
   not used to insert sorted or to find nodes

bee_tree_insert(..) will return NULL if memory allocation fails or if the
element is already in the tree.

If memory allocation fails errno will be set to ENOMEM.

If the element is already in the tree, errno will be set to EEXIST and the generated key
will be freed using the free_key() tree-functions (if set).