beesh: Add optional mee_check() target when --check is set
If the bee-file (beesh) is called with --check the special step "check" will
be called after calling "build" and before calling "install".

so if mee_check(), mee_check_pre() or mee_check_post() are defined they will be called
if --check is activated. As long as --check is not activated checks are disabled.

to globally enable checks you may set OPT_CHECK to any value but "no", e.g.
OPT_CHECK="yes".

we may rename those environment variables at a later time so please don't depend on them.
(it's just a quick fix - will be stable for bee-2.0)