beeversion: Add prefix char to print formatted extraname and extraversion
To prefix extraname and extraversion just add the prefix char before
e or x in the formatstring: e.g.

pkg_subname-1.0_extra1-0

%-e =>  -extra1
%_e =>  _extra1

%-x =>  -subname
%_x =>  _subname

This is only supposed to be used with '-' or '_' even if you may use other
chars here. But the behavior may change in future version. Only '_' and '-'
are stable for the moment.