bci: Ignore ENOTDIR for regular files while scanning directories
Scanning a directory should only fail if a CONTENT file is found
and we can't read this file for whatever reason or if we can't read
a directory that may contain a CONTENT file.

While scanning a directory ignoring regular files is fine.

mkdir a ; touch a/b ; bci a

should not fail with a/b/CONTENT: Not a directory