beesh: Allow bee-file to partially rename packages
Allow to prefix PKGNAME with 'something-'
Allow to append something to/set PKGEXTRANAME (e.g. '_something')
Allow to append something to/set PKGEXTRAVERSION (e.g. '_something')

In addition you may still set PKGARCH=any as before.

Changing anything else in the packagenames is currently not supported and
yields to an error message.

After renaming beeversion parses the new packagename again.

This is a cleaner version of the feature added in 3aaed0d