hooks: Fix gtk-update-icon-cache hooks (c84380db)
To check for a needed rebuild of the gtk icon cache database now
the existence of icons in the package is checked.

For all top level directories containing icons in some directory tree
the existence of an index.theme is checked. If it exists, we do call
gtk-update-icon-cache for this icon theme directory.

If no index.theme exists, we also delete the icon-theme.cache file.

Updating the icon cache is done in post-install and post-remove.

Since we do not yet have post-update post-install and post-remove will
be called for a package being updated and gtk-update-icon-cache will be
executed twice for all icon themes changed.

(cherry picked from commit c84380dbcbbe9e6f2c6d0550c2e1a4913f235db9)