beegetopt: Fix short option handling (again)
Fix in 678ce43c1b41c6def061a3f434b5071e04ccf6a4 was not well done.

We need to reset the unhandled short options so that option parsing may
continue after we hit a "-" invalid option.

We also need to decide if "-" is a special argument or an invalid option.