beesh: Add support for zip source archives
To emulate the strip-components behavior of tar we scan
the extracted sources and prepend a sourcesubdirectory
if the archive is extracted into a subdirectory.

This is a much cleaner solution then to strip components out of
pathnames within the archive.

So tar extraction should be changed too in the future.