buildtypes/cmake: Set `CMAKE_INSTALL_LIBDIR` to configured `LIBDIR`
It’s `lib64` by default [1].

> object code libraries (lib or lib64 or lib/<multiarch-tuple> on Debian)

This conflicts with the symbolic link in MarIuX.

    $ ls -ld /usr/lib*
    drwxr-xr-x 240 root root 143360 Feb 11 18:44 /usr/lib
    lrwxrwxrwx   1 root root      3 Mar  7  2011 /usr/lib64 -> lib
    drwxr-xr-x  21 root root   4096 Feb  5 16:45 /usr/libexec

Consider the current package has `/usr/lib/pkgconfig/libzip.pc`
installed, and the newer version installs the same file into
`/usr/lib64/pkgconfig/libzip.pc`. On upgrade, when removing the file
from the old package, bee will remove `/usr/lib64/pkgconfig/libzip.pc`
because of the symbolic link, which is unwanted.

So set it to `LIBDIR`, so that the system configuration is applied.

[1]: https://cmake.org/cmake/help/v3.8/module/GNUInstallDirs.html